apps/streamlit_apps/stoppages_analysis/requirements.txt
streamlit
pandas
plotly
Pillow
